requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=3.0
psycopg2-binary>=2.9
python-dateutil>=2.8.2
python-dotenv
boto3
ezdxf
matplotlib